requirements.txt
# Core ML
torch>=2.0.0
torchvision>=0.15.0

# Diffusion
diffusers>=0.25.0
transformers>=4.36.0
accelerate>=0.25.0

# CLIP
open-clip-torch>=2.24.0

# Optimization
cma>=3.3.0

# Utilities
Pillow>=10.0.0
numpy>=1.24.0
tqdm>=4.66.0
matplotlib>=3.8.0
scikit-learn>=1.3.0
scipy>=1.11.0

# For saving results
pandas>=2.0.0